fix(pgstress): enforce statement_timeout server-side and isolate it from admin work and the sampler
The statement timeout was a client-side context deadline. pgx answers an
expired context by closing the socket, so every timeout destroyed the pooled
connection: at a 3 ms timeout a 15 s run opened 12,434 new sessions and
completed zero transactions. With the timeout as a connection RuntimeParam
the server returns SQLSTATE 57014 on an intact session: 22 sessions and
8,291 transactions in the same run.

- NewPool takes a stmtTimeout and sets it in RuntimeParams; the client
  context becomes a long safety net (10x timeout + 30 s); classify maps
  57014 to "timeout"; maxConns is bounded before the int32 conversion.
- cmdRun opens a 2-connection admin pool without the timeout for in-run
  prepare/cleanup; the prepare and cleanup subcommands pass 0.
- The sampler runs on one dedicated *pgx.Conn cloned from the rw pool
  config with application_name=pgstress-sampler and its own
  statement_timeout equal to SampleInterval (a bare conn closed by a client
  deadline never recovers). It reconnects when closed, gives up on a view
  only for 42501/42P01/42883, drops an incomplete poll instead of publishing
  zeros, clamps rate, WAL and cache-hit deltas against counter resets,
  reports Deadlocks as a rate, retries server info after a failed first
  connect, and does not record cancellation at shutdown as an error.
- Startup refuses to run when MaxWorkers()+4 exceeds
  max_connections - superuser_reserved_connections on the rw host (and the
  ro host when it differs), printing both numbers and the remedies.
- clientInfo reports the rusage delta over the run rather than the process
  lifetime, so -prepare no longer inflates client CPU.

+++ stress/internal/metrics/types.go
@@ -81,7 +81,7 @@ type ServerSample struct {
 	TupModifiedPerSec float64         `json:"tupModifiedPerSec"`
 	TempBytesPerSec   float64         `json:"tempBytesPerSec"`
 	WALBytesPerSec    float64         `json:"walBytesPerSec"`
-	Deadlocks         int64           `json:"deadlocks"`
+	Deadlocks         float64         `json:"deadlocks"`
 	Active            int             `json:"active"`
 	Idle              int             `json:"idle"`
 	IdleInTx          int             `json:"idleInTx"`
